security: re-gate the media asset proxy behind the session (close the 2023 f6240fd2 regression)
`GET /legacy-api/ingest/recordings/:attr` proxies core media-api's
`/internal/assets/streams/:attr`, minting a machine-to-machine `systemUser`
Auth0 token per request. It has no guard of its own, and since commit f6240fd2
(2023-03-06, "don't use sessions for urls which use jwt") it has been mounted in
`non-session.js` -- i.e. BEFORE the session middleware. That commit moved 9
routes out of the session-gated router; 8 of them carry their own
`verifyToken()` + `hasRole([...])` guards and are called server-side by
rfcx-api, so the move was correct for them. This one relied on the session, and
moving it removed its only protection.

Because the proxy substitutes a systemUser token, media-api's per-user
`readableBy` scoping is skipped entirely (`user.has_system_role` -> readableBy
undefined -> no `hasPermission(READ, ...)` check). The route was therefore a
fully public, credential-free read of ANY stream in the platform.

Verified against production on 2026-08-10 before this change, with no cookie and
no Authorization header, from the public internet:
  - real 400x400 spectrogram PNG returned (200, 54,634 bytes)
  - raw audio for the same window: `_fwav.wav` 200 / 562,220 bytes (16-bit mono
    48kHz PCM), `_fmp3.mp3` 200 / 47,607 bytes
  - the sampled streams belong to a PRIVATE project (`is_public = f`)
  - served by ORIGIN, not just the CDN edge (`cf-cache-status: MISS`)
  - `checkAttrsValidity` allows windows up to 15 minutes, with arbitrary
    start/end/gain/frequency/dimensions -- so the exposure was the audio
    corpus, not a fixed set of thumbnails
Exposure window: ~2 years 5 months.

WHAT THIS COMMIT ADDS:
- app/routes/data-api/ingest-assets.js: the asset proxy, moved verbatim into its
  own router, with a header explaining why it must stay session-gated.

WHAT THIS COMMIT FIXES:
- app/routes/index.js: mounts `/legacy-api/ingest` (assets) AFTER the "Force
  login" middleware, restoring the pre-2023 protection.
- app/routes/data-api/ingest.js: drops the GET route (and the imports it alone
  used: `request`, `config`/`rfcxConfig`, `auth0Service`). The two JWT-guarded
  POSTs stay -- rfcx-api calls them server-side and must not need a session.
- app/routes/non-session.js: documents the invariant so an unguarded route is
  not added here again.
- Images are now `Cache-Control: private, max-age=31536000, immutable` rather
  than `public`. These responses are per-viewer authorised, so a SHARED cache
  entry could be served to an unauthenticated third party. `private` keeps the
  long-lived browser caching that makes the media-api render cache pay off.

WHY A COOKIE-BASED GATE WORKS FOR `<img>`:
A bare `<img src=...>` cannot send an `Authorization` header -- that is why the
proxy existed -- but it DOES send same-origin cookies. `parseTokenData()`
(app/middleware/jwt/index.js:58) falls back to `req.session.idToken` when no
Authorization header is present, and that is set at login (model/users.js). The
SPA is same-origin with legacy (arbimon.org) and already relies on this session
for `/legacy-api/*`.

VERIFIED:
- Routing-order harness reproducing the exact mount shape (non-session mounted
  before the session middleware, assets mounted after the login gate):
    POST /recordings/create  (no session) -> 200   [rfcx-api path intact]
    POST /recordings/delete  (no session) -> 200   [rfcx-api path intact]
    GET  /recordings/<attr>  anonymous    -> 302 /legacy-login
    GET  /recordings/<attr>  logged in    -> 200
- `node --check` clean on all four touched files.
- Non-browser caller audit: no server-side code fetches this path; all builders
  (`asset-url.js`, `recordings.js`, `models.js`, `classifications.js`,
  `templates.js`) emit RELATIVE URLs consumed by an authenticated browser, and
  the OpenGraph/resource-card pages that embed `recording.thumbnail` are
  themselves already session-gated (verified live: 302 anonymously). rfcx-api
  calls only the two guarded POSTs, never the GET.

OUT OF SCOPE / DEFERRED:
- Cloudflare purge: existing responses were cached `public, immutable`, so
  already-cached copies remain servable from the edge until purged. The purge
  must accompany the deploy.
- The direct stream-token route that ultimately retires this proxy (Track B) is
  a separate change in rfcx-local (public-router `/media-api/*`).
- The `/legacy-api` endpoints that are intentionally anonymous (`/alive`,
  `/projects-count`) are unchanged.
